switch to email password reset link (remove enter code screen)
