+++ frontend/src/app/(auth)/_layout.tsx
@@ -16,8 +16,8 @@ const AuthLayout = () => {
     <SafeAreaView style={{ flex: 1, backgroundColor: lightModeColors.background }}>
       <Stack screenOptions={{ headerShown: false }}>
         <Stack.Screen name="login" />
+        <Stack.Screen name="email-sent" />
         <Stack.Screen name="forgot-password" />
-        <Stack.Screen name="enter-code" />
       </Stack>
     </SafeAreaView>
   );
